requirements.txt
Flask==0.10.1
AfricastalkingGateway==1.8
